AgenticApproach/examples/requirements.txt
google-genai
langchain-google-genai
python-dotenv
httpx
tqdm
numpy<2
